Откорректировал основной объем приложения, начинаю писать тесты

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -1,11 +1,12 @@
 package ru.practicum.shareit.user;
 
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.exception.ConstraintViolationException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.shareit.exception.IncorrectParameterException;
 import ru.practicum.shareit.exception.ObjectNotFoundException;
 import ru.practicum.shareit.exception.ParameterException;
 import ru.practicum.shareit.user.dto.UserDto;
@@ -17,29 +18,37 @@ import static java.util.stream.Collectors.toList;
 @Slf4j
 @Service
 @Transactional(readOnly = true)
-@RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
+
+    @Autowired
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
     @Override
     @Transactional
     public UserDto save(UserDto userDto) {
+        valid(userDto);
         return UserMapper.toUserDto(userRepository.save(UserMapper.toUser(userDto)));
     }
 
     @Override
     @Transactional
     public UserDto update(UserDto userDto, Long id) {
-        User user;
-        user = userRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Пользователь с id = " + id + " не найден."));
-        if (userDto.getName() != null && !userDto.getName().isBlank()) {
+        User user = userRepository.findById(id).orElseThrow(() -> {
+            throw new ObjectNotFoundException("Пользователь с id = " + id + " не найден.");
+        });
+
+        if (userDto.getName() != null) {
             user.setName(userDto.getName());
         }
-        if (userDto.getEmail() != null && !userDto.getEmail().isBlank()) {
+        if (userDto.getEmail() != null) {
             user.setEmail(userDto.getEmail());
         }
+
         try {
-            return UserMapper.toUserDto(user);
+            return UserMapper.toUserDto(userRepository.save(user));
         } catch (DataIntegrityViolationException ex) {
             if (ex.getCause() instanceof ConstraintViolationException) {
                 throw new ParameterException(ex.getMessage());
@@ -50,13 +59,22 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto get(Long id) {
-        User user = userRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Пользователь с id = " + id + " не найден"));
+        if (id == null) {
+            throw new IncorrectParameterException("Id пользователя не может быть null");
+        }
+        User user = userRepository.findById(id).orElseThrow(() -> {
+            throw new ObjectNotFoundException("Пользователь с id = " + id + " не найден");
+        });
+
         return UserMapper.toUserDto(user);
     }
 
     @Override
     @Transactional
     public void delete(Long id) {
+        if (id == null) {
+            throw new IncorrectParameterException("Id не может быть пустым!");
+        }
         userRepository.deleteById(id);
     }
 
@@ -66,5 +84,29 @@ public class UserServiceImpl implements UserService {
                 .stream()
                 .map(UserMapper::toUserDto)
                 .collect(toList());
+    }
+
+    public void valid(UserDto user) {
+        if (user == null) {
+            throw new IncorrectParameterException("При создании пользователя передан некорреткный параметр");
+        } else if (user.getEmail() == null) {
+            throw new IncorrectParameterException("Email не может быть пустым");
+        } else if (!isValidEmailAddress(user.getEmail())) {
+            throw new IncorrectParameterException("Неверно задан email");
+        }
+    }
+
+    public boolean isValidEmailAddress(String email) {
+        String ePattern = "(?:[A-Za-z0-9!#$%&'*+/=?.^_`{|}~]" +
+                "+(?:.[a-z0-9!#$%&'*+/=?^_`{|}~]+)*|" +
+                "\\\"(?:[x01-x08x0bx0cx0e-x1fx21x23-x5bx5d-x7f]|[x01-x09x0bx0cx0e-x7f])*\\\")" +
+                "@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?.)" +
+                "+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|" +
+                "[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])).){3}(?:(2(5[0-5]|" +
+                "[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[x01-x08x0bx0cx0e-x1fx21-x5ax53-x7f]" +
+                "|[x01-x09x0bx0cx0e-x7f])+)])";
+        java.util.regex.Pattern p = java.util.regex.Pattern.compile(ePattern);
+        java.util.regex.Matcher m = p.matcher(email);
+        return m.matches();
     }
 }
